requirements.txt
eth-account==0.5.9
web3==5.31.0
python-dotenv==0.21.0
requests
statistics
